requirements.txt
streamlit
torch
torchvision
transformers
pillow
pandas
plotly
numpy
